+++ bundle/config/resources/model_serving_endpoint.go
@@ -2,6 +2,7 @@ package resources
 
 import (
 	"github.com/databricks/cli/bundle/config/paths"
+	"github.com/databricks/databricks-sdk-go/marshal"
 	"github.com/databricks/databricks-sdk-go/service/serving"
 )
 
@@ -14,11 +15,19 @@ type ModelServingEndpoint struct {
 	// as a reference in other resources. This value is returned by terraform.
 	ID string
 
-	// Local path where the bundle is defined. All bundle resources include
-	// this for interpolation purposes.
+	// Path to config file where the resource is defined. All bundle resources
+	// include this for interpolation purposes.
 	paths.Paths
 
 	// This is a resource agnostic implementation of permissions for ACLs.
 	// Implementation could be different based on the resource type.
 	Permissions []Permission `json:"permissions,omitempty"`
+}
+
+func (s *ModelServingEndpoint) UnmarshalJSON(b []byte) error {
+	return marshal.Unmarshal(b, s)
+}
+
+func (s ModelServingEndpoint) MarshalJSON() ([]byte, error) {
+	return marshal.Marshal(s)
 }