+++ bundle/internal/tf/schema/data_source_group.go
@@ -3,6 +3,7 @@
 package schema
 
 type DataSourceGroup struct {
+	AclPrincipalId          string   `json:"acl_principal_id,omitempty"`
 	AllowClusterCreate      bool     `json:"allow_cluster_create,omitempty"`
 	AllowInstancePoolCreate bool     `json:"allow_instance_pool_create,omitempty"`
 	ChildGroups             []string `json:"child_groups,omitempty"`