+++ cmd/account/metastore-assignments/metastore-assignments.go
@@ -280,7 +280,7 @@ func newList() *cobra.Command {
 
 		listReq.MetastoreId = args[0]
 
-		response, err := a.MetastoreAssignments.List(ctx, listReq)
+		response, err := a.MetastoreAssignments.ListAll(ctx, listReq)
 		if err != nil {
 			return err
 		}