+++ cmd/workspace/storage-credentials/storage-credentials.go
@@ -77,14 +77,7 @@ func newCreate() *cobra.Command {
 	cmd.Short = `Create a storage credential.`
 	cmd.Long = `Create a storage credential.
   
-  Creates a new storage credential. The request object is specific to the cloud:
-  
-  * **AwsIamRole** for AWS credentials. * **AzureServicePrincipal** for Azure
-  credentials. * **AzureManagedIdentity** for Azure managed credentials. *
-  **DatabricksGcpServiceAccount** for GCP managed credentials.
-  
-  The caller must be a metastore admin and have the
-  **CREATE_STORAGE_CREDENTIAL** privilege on the metastore.`
+  Creates a new storage credential.`
 
 	cmd.Annotations = make(map[string]string)
 
@@ -371,9 +364,7 @@ func newUpdate() *cobra.Command {
 	cmd.Short = `Update a credential.`
 	cmd.Long = `Update a credential.
   
-  Updates a storage credential on the metastore. The caller must be the owner of
-  the storage credential or a metastore admin. If the caller is a metastore
-  admin, only the __owner__ credential can be changed.`
+  Updates a storage credential on the metastore.`
 
 	cmd.Annotations = make(map[string]string)